Fixing implicit numeric widening warnings.

+++ app/uk/gov/hmrc/softdrinksindustrylevystub/services/SdilNumberTransformer.scala
@@ -43,8 +43,8 @@ object SdilNumberTransformer {
   def showTable(num: Int): Unit = {
     for {
       i       <- 1 to num
-      utr     <- tolerantUtr.get(i)
-      sdilRef <- sdilRefEnum.get(i)
+      utr     <- tolerantUtr.get(i.toLong)
+      sdilRef <- sdilRefEnum.get(i.toLong)
     } yield {
       println(s"| $utr | $sdilRef |")
     }
